pyproject.toml
[project]
name = "manga-colorization"
version = "1.0.0"
description = "Pix2Pix GAN for automatic manga colorization"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "matplotlib>=3.7.0",
    "pandas>=2.0.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "pytest>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
